refactor(tests): split unit and integration test suites
