+++ phpunit.xml
@@ -10,11 +10,14 @@
         beStrictAboutCoverageMetadata="false"
 >
     <testsuites>
-        <testsuite name="default">
-            <directory suffix="Test.php">tests</directory>
+        <testsuite name="unit">
+            <directory suffix="Test.php">tests/Unit</directory>
+        </testsuite>
+
+        <testsuite name="integration">
+            <directory suffix="Test.php">tests/Integration</directory>
         </testsuite>
     </testsuites>
-
 
     <source>
         <include>
